Extract `AnvilCompilation` into its own file.
